Use Bundler to automatically verify list of dependencies
We want to make sure that all dependencies are properly declared in our
agraph.gemspec file.  The easiest way to ensure this is to force our
specs to fail if we accidentally depend on an unlisted gem.

By using bundler, we can precisely control which gems are available in
our search path, and we can force rubygems to raise an error if we
depend on something that we haven't declared.

To prevent code duplication, we use a magic code snippet in our Gemfile
which reads the existing agraph.gemspec and automatically declares the
gems listed there in appropriate groups.

We also needed to add 'rake' as an explicit development dependency.

To set up a new development system, install the latest version of the
'bundler' gem, and run 'bundle install'.

The accompanying Gemfile.lock provides a list of exact gem versions
known to work.  It can be updated at any time using the command 'bundle
update'.